backend/requirements.txt
websockets==12.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
pygeohash==1.2.0
aiohttp==3.9.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dateutil==2.9.0.post0
pip-system-certs==5.3.0

